fix: added track requests handler
