+++ src/pages/TrackInitiatedRequests/TrackInitiatedRequests.tsx
@@ -2,13 +2,14 @@ import React from 'react'
 import InnerBodyHeader from '../../components/InnerBodyHeader/InnerBodyHeader'
 import styles from './TrackInitiatedRequests.module.css'
 import TrackRequestsContainer from '../../components/TrackRequestsContainer/TrackRequestsContainer'
+import TrackRequestsContainerHandler from '../../components/TrackRequestsContainer/TrackRequestsContainerHandler'
 
 const TrackInitiatedRequests = () => {
   return (
     <>
       <div className={styles.main}>
         <InnerBodyHeader heading="Track Requests" />
-        <TrackRequestsContainer/>
+        <TrackRequestsContainerHandler/>
       </div>
     </>
   )
